lint: enable forcetypeassert, unconvert and unused
The three were commented out; enabling them reported five issues, all
redundant conversions or an unchecked assertion in tests, fixed here.

+++ internal/output/table.go
@@ -111,13 +111,13 @@ func HumanDuration(d time.Duration) string {
 	} else if hours < 24*365*2 {
 		return fmt.Sprintf("%dd", hours/24)
 	} else if hours < 24*365*8 {
-		dy := int(hours/24) % 365
+		dy := hours / 24 % 365
 		if dy == 0 {
 			return fmt.Sprintf("%dy", hours/24/365)
 		}
 		return fmt.Sprintf("%dy%dd", hours/24/365, dy)
 	}
-	return fmt.Sprintf("%dy", int(hours/24/365))
+	return fmt.Sprintf("%dy", hours/24/365)
 }
 
 // FormatLabels returns a comma-separated key=value string from a label map,
